+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+ - EC2 instance state attacks, i.e., EC2 instance stop, reboot, hibernate and termiante.
+
 ## 1.0.0
 
  - Initial release